tests/wpt/tests/css/css-display/parsing/tentative/display-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Display: Parsing display with valid values</title>
<link rel="author" title="Ethan Jimenez" href="mailto:ethavar@microsoft.com">
<link rel="help" href="https://tabatkins.github.io/specs/css-masonry/#masonry-containers">
<meta name="assert" content="display supports the new values 'grid-lanes | inline-grid-lanes'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
// https://tabatkins.github.io/specs/css-masonry/#masonry-containers
test_valid_value("display", "grid-lanes");
test_valid_value("display", "inline-grid-lanes");
</script>
</body>
</html>
